requirements.txt

numpy
pandas
scikit-learn
pillow
openpyxl
python-docx
python-pptx
docx2txt
PyPDF2
pytesseract

sentence-transformers>=2.2.0

plotly>=5.0.0
pyvis>=0.3.1

weaviate-client>=4.7.0
requests
validators
authlib
grpcio
grpcio-tools
grpcio-health-checking
httpx
tqdm

